error

+++ src/software/jetson_nano/thunderloop.cpp
@@ -117,7 +117,8 @@ Thunderloop::~Thunderloop() {}
                 new_world         = std::get<1>(result);
             }
 
-            thunderloop_status_.set_network_service_poll_time_ms(getNanoseconds(poll_time));
+            thunderloop_status_.set_network_service_poll_time_ms(
+                getNanoseconds(poll_time));
 
             network_status_.set_ms_since_last_primitive_received(UINT32_MAX);
             network_status_.set_ms_since_last_vision_received(UINT32_MAX);
@@ -141,7 +142,7 @@ Thunderloop::~Thunderloop() {}
                         &time_since_last_primitive_received);
                     clock_gettime(CLOCK_MONOTONIC, &last_primitive_received_time);
                     network_status_.set_ms_since_last_primitive_received(
-                            getNanoseconds(time_since_last_primitive_received));
+                        getNanoseconds(time_since_last_primitive_received));
 
                     // Start new primitive
                     {
@@ -149,7 +150,8 @@ Thunderloop::~Thunderloop() {}
                         primitive_executor_.updatePrimitiveSet(primitive_set_);
                     }
 
-                    thunderloop_status_.set_primitive_executor_start_time_ms(getNanoseconds(poll_time));
+                    thunderloop_status_.set_primitive_executor_start_time_ms(
+                        getNanoseconds(poll_time));
                 }
             }
 
@@ -163,7 +165,8 @@ Thunderloop::~Thunderloop() {}
                                                   &last_world_recieved_time,
                                                   &time_since_last_vision_received);
                 clock_gettime(CLOCK_MONOTONIC, &last_world_recieved_time);
-                network_status_.set_ms_since_last_vision_received(getNanoseconds(time_since_last_vision_received));
+                network_status_.set_ms_since_last_vision_received(
+                    getNanoseconds(time_since_last_vision_received));
 
                 primitive_executor_.updateWorld(new_world);
                 world_ = new_world;
@@ -175,10 +178,11 @@ Thunderloop::~Thunderloop() {}
                 ScopedTimespecTimer timer(&poll_time);
 
                 // Handle emergency stop override
-                auto nanoseconds_elapsed_since_last_primitive = getNanoseconds(time_since_last_primitive_received);
+                auto nanoseconds_elapsed_since_last_primitive =
+                    getNanoseconds(time_since_last_primitive_received);
 
                 if (nanoseconds_elapsed_since_last_primitive >
-                    static_cast<long>(PRIMITIVE_MANAGER_TIMEOUT_NS))
+                    PRIMITIVE_MANAGER_TIMEOUT_NS)
                 {
                     primitive_executor_.setStopPrimitive();
 
@@ -193,7 +197,8 @@ Thunderloop::~Thunderloop() {}
                 direct_control_ = *primitive_executor_.stepPrimitive();
             }
 
-            thunderloop_status_.set_primitive_executor_step_time_ms(getNanoseconds(poll_time));
+            thunderloop_status_.set_primitive_executor_step_time_ms(
+                getNanoseconds(poll_time));
 
             // Power Service: execute the power control command
             {
@@ -214,7 +219,8 @@ Thunderloop::~Thunderloop() {}
                 ScopedTimespecTimer::timespecDiff(&current_time, &last_kicker_fired,
                                                   &time_since_kicker_fired);
                 clock_gettime(CLOCK_MONOTONIC, &last_kicker_fired);
-                chipper_kicker_status_.set_ms_since_kicker_fired(getNanoseconds(time_since_kicker_fired));
+                chipper_kicker_status_.set_ms_since_kicker_fired(
+                    getNanoseconds(time_since_kicker_fired));
             }
             else if (direct_control_.power_control().chicker().has_chip_distance_meters())
             {
@@ -223,7 +229,8 @@ Thunderloop::~Thunderloop() {}
                 ScopedTimespecTimer::timespecDiff(&current_time, &last_chipper_fired,
                                                   &time_since_chipper_fired);
                 clock_gettime(CLOCK_MONOTONIC, &last_chipper_fired);
-                chipper_kicker_status_.set_ms_since_kicker_fired(getNanoseconds(time_since_chipper_fired));
+                chipper_kicker_status_.set_ms_since_kicker_fired(
+                    getNanoseconds(time_since_chipper_fired));
             }
 
             // Motor Service: execute the motor control command
@@ -238,7 +245,8 @@ Thunderloop::~Thunderloop() {}
             thunderloop_status_.set_motor_service_poll_time_ms(getNanoseconds(poll_time));
 
             clock_gettime(CLOCK_MONOTONIC, &current_time);
-            time_sent_.set_epoch_timestamp_seconds(static_cast<double>(current_time.tv_sec));
+            time_sent_.set_epoch_timestamp_seconds(
+                static_cast<double>(current_time.tv_sec));
 
             // Update Robot Status with poll responses
             robot_status_.set_robot_id(robot_id_);
@@ -260,14 +268,15 @@ Thunderloop::~Thunderloop() {}
         }
 
         auto loop_duration = getNanoseconds(iteration_time);
-        thunderloop_status_.set_iteration_time_ms(loop_duration / NANOSECONDS_PER_MILLISECOND);
+        thunderloop_status_.set_iteration_time_ms(loop_duration /
+                                                  NANOSECONDS_PER_MILLISECOND);
 
         // Make sure the iteration can fit inside the period of the loop
         loop_duration_seconds =
             static_cast<double>(loop_duration) * SECONDS_PER_NANOSECOND;
 
         // Calculate next shot taking into account how long this iteration took
-        next_shot.tv_nsec += interval - loop_duration;
+        next_shot.tv_nsec += interval - static_cast<long int>(loop_duration);
         timespecNorm(next_shot);
     }
 }
@@ -275,7 +284,7 @@ Thunderloop::~Thunderloop() {}
 double Thunderloop::getNanoseconds(timespec time)
 {
     return (static_cast<double>(time.tv_nsec) * 1000) +
-            (static_cast<double>(time.tv_nsec) / NANOSECONDS_PER_MILLISECOND);
+           (static_cast<double>(time.tv_nsec) / NANOSECONDS_PER_MILLISECOND);
 }
 
 void Thunderloop::timespecNorm(struct timespec& ts)
